Replace cache keys with md5(key)
The memcached client limits the length of keys to 250 characters, so we
can't always use the full URL.  This patch changes it to use md5(key)
instead.  We assume there will not be any collisions.  This is what
graphite-web is doing as well.

On the off chance this becomes a bottleneck, we can replace
md5 with one or two calls to siphash, and sprintf() with something that
just does hexstring conversion.