Add movingAverage() that processes previous points
The previous movingAverage() implemenentation started averaging at the
start of the time series.  This implementation begins *before* the time
series that the first point is already the moving average of the
preceding n points.

Note that this currently only works correctly with secondly metrics or
by providing a time-based offset.