Revert to old-style graphite behaviour for movingAverage()
Previously we attempted to timeshift the data so that the first data
point was truly the average of the previous (out-of-window) data points.
However, this only sort of worked.  For one, it assumed the metrics were
all stored at second resolution.  This is not possible to work around
with sending a second query to the Zipper.

Now instead, elements before the moving average window is filled will be
None.