Fix rendering bug when data stream contains nulls
When we have more points than pixels, we average out groups of points
to figure out how to plot it.  If the data series contains nulls, we
were incorrectly including them in the average resulting in sharp drops
in the graph whenever they were present.

This patch ignores nulls from the average calculation as graphite does.