holtWinters forecasting
Add support for holtWinters forecasting. Behaviour is improvement on
graphite behavior, so we are breaking compatibility.

Bootstrap + original data is retrieved in a single request with a single
consistent stepTime, unlike what graphite does, which is to retrieve two
datasets, and then it linearly amplifies (i.e., multiplies) the
bootstrap by a factor to match the granularity of the original
dataset. It ignores the aggregation function that might have been
applied to do the rollup, in carbonapi we respect it.