move functions and helpers, expr parsing into their own files for clarity
also:
* use some closures to more elegantly reuse configurable "base" functions
* ignore carbonapi binary
* currentValue to curValue for consistency

the expr file was a bit of a mess. this is a first step in bringing in
some sanity.
I'ld like to more cleanly separate the expression parsing from the
execution (functions), right now the functions are still a bit too
closely tied to the expr objects, and still do some parsing (e.g.
arguments) on their own.
Also many of the structs are used everywhere (data rendering, expr
parsing, function execution), such as metricData, metricRequest, in
addition to the already mentioned expr type, this is holding back
in further isolating responsabilities.  Maybe that will be a next
step...