Add smoothedMovingAverage / smma(seriesList)
An exponentially weighted moving average (see: `ewma(seriesList, alpha)`) where the alpha
component is calculated as 1 / ( n : the number of defined points in
the series ).

--
I had a private branch of ewma and smma from back in December that I
never merged. This is the smma portion, lightly translated to use the
same library as we did for ewma.