Include the search backend in the Limiter servers.
The search backend was subject to the same Limiter behavior as
everything else fetched through singleGet, but since it wasn't present
in the Limiter set of servers, the limit was effectively 0, and so zero
queries were allowed through to the search backend.

This was missed during testing because ConcurrencyLimitPerServer was set
to 0, which disables the Limiter behavior.