add expr lowPass to use with fft and ifft
Once we use 'fft' there are several operations we can perform on its
result before converting it back to a time series with 'ifft'.

'removeBelowValue' is one example, but another important function is
the low pass filter. This commit is a version of such a filter that
can operate directly on the output of the 'fft' as follows:

    ifft(lowPass(fft(metric1,'abs'),0.2),fft(metric1,'phase'))