Introduce SendGlobsAsIs configuration flag
This will allow you to turn on or off the behavior that was introduced in 0.7.0

Apparantly if you'll send requests as is, memory consumption on carbonzipper
side will significantly increase. This switch will allow users to switch
this behavior on or off, depends on what they need.

This fix should be considered as temporary until we implement better solution.