expr diffSeries: allow subtracting series with different StepTimes
when you subtract series with different step times in carbonapi
or python graphite you'll get an exception.

but it's usefult to assume that, if the subtrahend step time is a
multiple of the minuend step time, the minuend is the average value
for a range.

this can be used on comparing data that has some seasonality.

    diffSeries(some.metric, summarize(some.metric, 'period', 'avg'))

the result will be that every period of that will be normalized
to the same average.

I expect to use this result with ifft(fft()) so we can create
predictions.