Fix timeShift bug
As described in the previous commit, there is a bug in timeShift that
makes it return no metrics no matter what we pass in. The problem is
that we define

    fromNew := int64(int64(from) + int64(offs))
    untilNew := int64(int64(from) + int64(offs))

and thus fromNew == untilNew.

Most of these type casts are no longer necessary, so this patch gets rid
of them and moves the computation of the new from/until parameters into
a method call.