Prefer from/until times in response
Version 3 of the protocol now includes the original request start- and
end times for each metric in the responses. Using this is more reliable
than trying to match metric names to requests.

When combined with a pending pull-request to lomik/go-carbon, this
commit fixes issue 319 _for version 3 of the protocol_:

    https://github.com/go-graphite/carbonapi/issues/319
    https://github.com/lomik/go-carbon/pull/238

Version 2 of the protocol still suffers from issue 319, for different
reasons.