Correctly batch metric fetching in protocol v2
When we fetch metrics in protocol v2, we use URL parameters to send the
information of what metrics and time periods we want to the backends.
When we do this, we can group metric targets together by path expression
to reduce the number of round-trips we have to make to the backends.

However, due to functions like timeShift and timeStack, we must also
group metrics together according to the start- and end times we want to
fetch. This patch acheives that, and fixes issue 63 in Github:

    https://github.com/go-graphite/carbonzipper/issues/63