Remove protobuf definitions from access log
We weren't serializing the access log entries as protobuf, and I'm not
sure in what situation we'd want to do that. This commit leaves only the
JSON serialization code in place. It can be edited independently, which
we want to do to remove some `omitempty` declarations so our
ElasticSearch queries can become more useful.