Make no-reply errors less spammy
We see _a lot_ of these no reply errors in production. Their root cause
might be that we fan out requests to all our clusters, irrespective of
whether we suspect those clusters contain the metrics we're asking for
or not. We're looking into that, but until then we'd like a little less
spam in our logs and monitoring.