Flatten group tree before sending requests
While the new config format is tree-like, and we marshal it into a tree
structure in our code, it's not necessary to respect that tree structure
when we make our Fetch() calls.

Instead, we can start the BroadcastGroup.Fetch() call by recursively
linearizing the tree into a simple list of ServerClient structures, and
make a concrete Fetch() call on each leaf node directly. This has some
benefits over resolving the Fetch() calls themselves recursively:

(1) This actually works. We saw a lot of problems with the old approach
in our production tests, where requests took much longer than in the
legacy code.

(2) The call stack of a Fetch() call is simpler, which is nice for
reasoning about the behaviour of the code, as well as when stepping
through it in a debugger.

(3) It's relatively clear how to implement pre-call filtering logic on
top of a list of ServerClient structs in the future.