Simplify fetch response merging
We only handle the case where the metrics have the same start time and
the same step size. This is what carbonzipper currently does in
production.

Later we'll want to do something about the case where the metrics have
differing step sizes, but we need to gather some data on cases where
that happens before we can make a decision about what to do. We add some
logging to help with this.