Delete broken test
I haven't bisected to find out for sure, but this test was probably
broken by commit 8a3ec301. We're going to delete it because it's not
clear what we should do in the case when we get two responses for the
same metric with different resolutions.

We may eventually decide to downsample the metric of higher resolution,
but it's been pointed out that we should maybe take into account any
user-specified summarize calls before we do that. The reason is that
users have already been trained to deal with this case, and the ones
that cared to take the time to do so probably thought about how they
wanted to do that, so we should preserve that behaviour.