Fix tag parsing for various replies from backends
Backends in some cases might send responses with tags that are not valid
according to https://graphite.readthedocs.io/en/latest/tags.html

In that case we want to do our best to handle such replies.

This commit changes current behavior (panic on any invalid reply and truncate other replies)
to be more sane but not necessary match graphite's.

Beahvior change are:
1. Ignore empty tags (cut `;` from the tag if it was present), e.x. `metric1;;tag`
2. Proper handling of values that have `=` in them
3. If tag doesn't have value - consider that an empty value, e.x. `metric1;tag`
4. Tags might have empty values e.x. `metric1;tag1=;tag2=v2`

Fixes #565