transformNull: add defaultOnAbsent argument
This change is based on idea of function `event` from avito's carbonapi fork.

Original code: https://github.com/el-yurchito/carbonapi/tree/master/expr/functions/event
Original authors: @PAFomin-at-avito, @el-yurchito

This adds an argument to transformNull function that would create a constant metric if
no results were returned from a backend.

This might be useful in case you are working with events that might not be present
for a certain periods of time in a past.