Port heatMap function from avito's carbonapi fork
Original code: https://github.com/el-yurchito/carbonapi/tree/master/expr/functions/heatMap
Original authors: @el-yurchito

This is a heat-map like function that do diff between series, based on their computed weight.

All metrics are assigned weights, based on the sum of their first 5 values and then sorted based on that.

After that for the sorted metrics, diff with the previous one will be computed.

Assuming seriesList has values N series in total (sorted by sum of the first 5 values):
(a[1], a[2], ..., a[N]). Then heatMap will output N-1 series: (a[2] - a[1], a[3] - a[2], ..., a[N] - a[N-1]).