Update Description definition to be able to handle same parameters as graphite-web@master
In current master, there are some new argument types and options available,

e.x. for smartSummarize and option `alignTo` it is not possible to specify
`null`, `true` and `false` as well as `"minutes"`, `"years"`, etc.

This requires some changes to functiondiff tool and to all to do some
internal data-type changes for the descriptions struct.