Fix handling of partial-failed fetch requests that contain globs (only when maxBatchSize > 0)
When request splitting was used and some of the servers didn't replied
but others did - such requests were handled as failed and no data
was returned, which resulted in error 500 from carbonapi.

Changing behavior in that case so requests that went through splitting
procedure will be fetched despite of partial failure to expand globs.

Fixes #545