Add dependabot configuration for golang dependencies
This commits add configuration for the github depedanbot
to check weekly for newer versions of the golang dependencies
which are used in this project.
This should help with maintenance, since it is easier to
do some occasional update with small changes then doing a big one
when there is a known security problem or something like
that.